+++ impls/java23/src/main/java/mal/Env.java
@@ -8,6 +8,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import mal.Mal.MalIterable;
 import mal.Mal.MalSymbol;
 
 public class Env {
@@ -31,7 +32,7 @@ public class Env {
     this(outer, new HashMap<>());
   }
 
-  public Env(Env outer, Iterable<Mal> binds, Iterable<Mal> exprs) {
+  public Env(Env outer, MalIterable binds, MalIterable exprs) {
     this(outer, toMap(binds, exprs));
   }
 
@@ -60,7 +61,7 @@ public class Env {
     map.put(key.name(), value);
   }
     
-  private static Map<String, Mal> toMap(Iterable<Mal> binds, Iterable<Mal> exprs) {
+  private static Map<String, Mal> toMap(MalIterable binds, MalIterable exprs) {
     var i = binds.iterator();
     var j = exprs.iterator();
